add DType type alias
