+++ spec/API_specification/dataframe_api/dtypes.py
@@ -1,0 +1,36 @@
+##########
+# Dtypes #
+##########
+
+class Int64:
+    """Integer type with 64 bits of precision."""
+
+class Int32:
+    """Integer type with 32 bits of precision."""
+
+class Int16:
+    """Integer type with 16 bits of precision."""
+
+class Int8:
+    """Integer type with 8 bits of precision."""
+
+class UInt64:
+    """Unsigned integer type with 64 bits of precision."""
+
+class UInt32:
+    """Unsigned integer type with 32 bits of precision."""
+
+class UInt16:
+    """Unsigned integer type with 16 bits of precision."""
+
+class UInt8:
+    """Unsigned integer type with 8 bits of precision."""
+
+class Float64:
+    """Floating point type with 64 bits of precision."""
+
+class Float32:
+    """Floating point type with 32 bits of precision."""
+
+class Bool:
+    """Boolean type with 8 bits of precision."""
